Update Readme and some refactor
